backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic-settings
bcrypt
python-jose[cryptography]
email-validator
python-multipart
passlib